Yoga/requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.1
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
DateTime==4.3
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==3.1.1
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-extensions==3.1.0
django-import==0.0.4
django-import-export==2.4.0
django-js-asset==1.2.2
django-jsoneditor==0.1.6
djangorestframework==3.12.0
et-xmlfile==1.0.1
export==0.1.2
filelock==3.0.12
future==0.18.2
html5lib==1.1
idna==2.10
importlib-metadata==2.0.0
isort==5.5.1
jdcal==1.4.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupPy==1.14
matplotlib==3.3.3
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.21

odfpy==1.4.1
openpyxl==3.0.5
packaging==20.7
panda==0.3.1
pandas==1.1.3

protobuf==3.13.0
pylint==2.6.0
pymongo==3.11.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
razorpay==1.2.0
reportlab==3.5.55
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
virtualenv==20.0.35
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
zipp==3.3.1
zope.interface==5.2.0
